reorganize trainer
